fix: correções eduardo conversa

+++ os/memory_manager.cpp
@@ -10,7 +10,7 @@ namespace OS {
     MemoryManager* memory_manager = nullptr;
 
     MemoryManager::MemoryManager(const uint16_t size) : next_free_addr(0), PHYSICAL_MEMORY_SIZE(size) {
-        uint32_t PROXIMO_PAGE = 512;
+        uint16_t PROXIMO_PAGE = 512;
         next_free_addr = PROXIMO_PAGE;
     }
